+++ drivers/gpu/nvgpu/hal/fb/fb_gp10b.c
@@ -1,7 +1,7 @@
 /*
  * GP10B FB
  *
- * Copyright (c) 2014-2018, NVIDIA CORPORATION.  All rights reserved.
+ * Copyright (c) 2014-2019, NVIDIA CORPORATION.  All rights reserved.
 *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -27,7 +27,7 @@
 
 #include "fb_gp10b.h"
 
-unsigned int gp10b_fb_compression_page_size(struct gk20a *g)
+u64 gp10b_fb_compression_page_size(struct gk20a *g)
 {
 	return SZ_64K;
 }
